feat: Add route for generating QRCode for a file
Add a new route `/qrcode/:id` to generate a QRCode image for a file. This route will provide a link to the file in the form of a QRCode image. This feature enhances the functionality of the file management system by allowing users to easily access files using QRCode scanning.